Update movement.c
This is what I've got so far for using the cliff sensor. It may need some adjusting. I set the value in the second half to 4095 because that was the max of the range that was listed on the OI spec sheet. It would probably be better to have it be smaller